requirements.txt
tensorflow-gpu
matplotlib
librosa
scipy
sklearn
pandas
numpy
tqdm
os

